Log processed resources in output-only mode
There are some things we do to the resources when actually pushing them into
kubernetes, and just dumping the _rendered template_ is not enough for debugging
through that.

Note that in the future we could actually move the writing of the rendered template to
here as well, simplifying the code: We do not _need_ the rendered template as a file.