Fix the implementation for the "include modules" capability
When the user provides modules explicitly on the command line, then we want
at most use these modules.
If a module specified by the user as "included" doesn't actually exist, then
we need to flag an error.
If a module specified by the user as "included" is also specified as "excluded",
then we similarly need to flag an error.
If a module specified by the user as "excluded" is not actually available, then
that is _not_ an error: We will certainly not include it.