Make resource application more flexible
Patching has a lot of issues, as it would lead to "drift" of the actual state compared
to the desired state as documented by the templates.
Updating on the other hand also has issues, for example with services (which may have
a auto-updated 'clusterIP' field) or configmaps/secrets (which may be used to keep state
in the environment that is not in the templates -- secrets!)

This commit introduces a 'bootstrap.k8s.collaborne.com/strategy' annotation, which contains
a list of verbs that describe how to update a specific resource:

* 'create', 'update', 'patch': Try these operations.
  'patch' is somewhat smart in trying with both strategic and regular merge patch approaches
* 'fail', 'skip': Fail application, or skip application without error
* 'smart': Use patch for the mentioned service/configmaps/secret resources, and update for anything
  else. If that doesn't work because the resource doesn't exist: Try to create it.
* 'legacy': The existing behavior (with minimal code changes needed for the refactoring)

The default strategy is 'legacy' for compatibility, but the idea is to change that default
eventually.

Note that the other annotations are merged into the strategy behavior:
* 'update-allowed':
  - Respected by the legacy strategy unchanged
  - Respected by the smart strategy to switch to patching for resources that would typically get updated.
  - Leads to an error if the strategy is explicitly configured and contains the verb 'update'
* 'ignore-patch-failures':
  - Respected by the legacy strategy unchanged
  - Translated as implicit 'skip' if the strategy is explicitly configured

Fixes: #8