Convert to async/await
This makes the applyResource() function "bigger", but that actually just shows the
complexity of the current approach better. Future commits will start chipping away
on that again.

Note that this also kills the logResult function, which only existed to make _writing_
the `then`s a bit easier, but bought this with a considerable increase in mental
effort required to _read_ them.

As a side-effect this commit fixes the "outer" error logging to report the last operation
attempted, rather than logging `(undefined)`.