Clarify the comment in stategyFail
Initially the idea was to add a 'create' strategy as a default behavior, and therefore
then we needed a way to prevent that. But, the strategy logic has evolved quite a bit, and
it was found that only having that default inside of the complex 'smart'/'legacy' strategies
worked better.

Right now using 'fail' as a strategy verb is pretty much a no-op.